+++ components/compositing/Cargo.toml
@@ -22,6 +22,9 @@ path = "../msg"
 [dependencies.net]
 path = "../net"
 
+[dependencies.profile]
+path = "../profile"
+
 [dependencies.util]
 path = "../util"
