+++ components/compositing/constellation.rs
@@ -25,12 +25,12 @@ use msg::constellation_msg::Msg as ConstellationMsg;
 use net::image_cache_task::{ImageCacheTask, ImageCacheTaskClient};
 use net::resource_task::{self, ResourceTask};
 use net::storage_task::{StorageTask, StorageTaskMsg};
+use profile::mem::MemoryProfilerChan;
+use profile::time::TimeProfilerChan;
 use util::cursor::Cursor;
 use util::geometry::PagePx;
-use util::memory::MemoryProfilerChan;
 use util::opts;
 use util::task::spawn_named;
-use util::time::TimeProfilerChan;
 use std::borrow::ToOwned;
 use std::collections::HashMap;
 use std::old_io as io;
