+++ components/gfx/paint_task.rs
@@ -25,17 +25,17 @@ use msg::compositor_msg::{LayerMetadata, PaintListener, ScrollPolicy};
 use msg::constellation_msg::Msg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use msg::constellation_msg::PipelineExitType;
+use profile::time::{TimeProfilerChan, TimeProfilerCategory, profile};
 use skia::SkiaGrGLNativeContextRef;
+use std::mem;
+use std::thread::Builder;
+use std::sync::Arc;
+use std::sync::mpsc::{Receiver, Sender, channel};
 use util::geometry::{Au, ZERO_POINT};
 use util::opts;
 use util::smallvec::SmallVec;
 use util::task::spawn_named_with_send_on_failure;
 use util::task_state;
-use util::time::{TimeProfilerChan, TimeProfilerCategory, profile};
-use std::mem;
-use std::thread::Builder;
-use std::sync::Arc;
-use std::sync::mpsc::{Receiver, Sender, channel};
 
 /// Information about a hardware graphics layer that layout sends to the painting task.
 #[derive(Clone)]
