+++ components/layout/Cargo.toml
@@ -37,6 +37,9 @@ path = "../plugins"
 [dependencies.net]
 path = "../net"
 
+[dependencies.profile]
+path = "../profile"
+
 [dependencies.util]
 path = "../util"
