+++ components/layout_traits/Cargo.toml
@@ -19,6 +19,9 @@ path = "../msg"
 [dependencies.net]
 path = "../net"
 
+[dependencies.profile]
+path = "../profile"
+
 [dependencies.util]
 path = "../util"
