+++ components/net/Cargo.toml
@@ -7,6 +7,9 @@ authors = ["The Servo Project Developers"]
 name = "net"
 path = "lib.rs"
 
+[dependencies.profile]
+path = "../profile"
+
 [dependencies.util]
 path = "../util"
