+++ components/net/lib.rs
@@ -26,6 +26,7 @@ extern crate png;
 #[macro_use]
 extern crate log;
 extern crate openssl;
+extern crate profile;
 extern crate "rustc-serialize" as rustc_serialize;
 extern crate util;
 extern crate stb_image;
