+++ components/script/lib.rs
@@ -43,6 +43,7 @@ extern crate net;
 extern crate "rustc-serialize" as rustc_serialize;
 extern crate time;
 extern crate canvas;
+extern crate profile;
 extern crate script_traits;
 extern crate selectors;
 extern crate util;
